Mode validation done

+++ S148.Backend.Shopping.Service/Repositories/IProductRepository.cs
@@ -4,5 +4,5 @@ public interface IProductRepository
 {
     IReadOnlyCollection<int> GetAll();
 
-    float? GetPrice(int id);
+    decimal? GetPrice(int id);
 }
